build upgrade system and use f32

+++ src/demo/gameplay/dust.rs
@@ -26,7 +26,7 @@ struct Velocity {
 }
 
 // TODO: impl health component for dust
-pub const DUST_HEALTH: u32 = 5;
+pub const DUST_HEALTH: f32 = 5.0;
 
 pub fn dust(pos: Vec2, speed: f32) -> impl Bundle {
     (
